pyproject.toml
[tool.poetry]
name = "pydantic-settings"
version = "0.1.2b"
description = "Hipster-orgazmic tool to mange application settings"
readme = "README.md"
authors = ["Daniel Daniels <danields761@gmail.com>"]
license = "MIT"
repository = "https://github.com/danields761/pydantic-settings"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: MIT License",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Internet",
]

[tool.poetry.dependencies]
python = "^3.6"
pydantic = "^0.32.*"
pyyaml = "^5.1.2"
tomlkit = "^0.5.5"
typing-extensions = "^3.7.4"
attrs = "^19.1"
class-doc = "0.2.0b"

[tool.poetry.dev-dependencies]
pytest = "^5.0.1"
black = "^19.3b0"
pytest-clarity = "^0.2.0-alpha.1"
sphinx = "^2.2.0"
sphinxcontrib-golangdomain = "^0.2.0-alpha.0"
sphinx-autoapi = "^1.1.0"
recommonmark = "^0.6.0"
m2r = "^0.2.1"

[tool.black]
# UEAH thats works
skip_string_normalization = true
exclude = "(docs/examples/*)|(\\.?venv)"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
